chore: remove tiny-glob dependency
