+++ packages/kit/package.json
@@ -22,7 +22,6 @@
 		"sade": "^1.8.1",
 		"set-cookie-parser": "^2.6.0",
 		"sirv": "^2.0.2",
-		"tiny-glob": "^0.2.9",
 		"undici": "~5.22.0"
 	},
 	"devDependencies": {
